structure the code
